Added line height support

+++ Src/Core/Headers/Texture.hpp
@@ -8,7 +8,7 @@ namespace AnA
     {
     public:
         Texture(const char* filename, Device& mDevice);
-        Texture(const char* text, const int width, const int height, const int lineHeight, Device& mDevice);
+        Texture(const char* text, const int width, const int height, const float lineHeight, Device& mDevice);
         ~Texture();
 
         VkImageView& GetImageView();
